requirements.txt
python-dotenv
weaviate-client
transformers
torch
pillow
numpy
openai
mcp[cli]
